+++ pkg/streamingpromql/engine_test.go
@@ -181,6 +181,7 @@ func TestNewInstantQuery_Strings(t *testing.T) {
 // Once the streaming engine supports all PromQL features exercised by Prometheus' test cases, we can remove these files and instead call promql.RunBuiltinTests here instead.
 func TestUpstreamTestCases(t *testing.T) {
 	opts := NewTestEngineOpts()
+	opts.CommonOpts.EnableDelayedNameRemoval = true
 	planner, err := NewQueryPlanner(opts)
 	require.NoError(t, err)
 	engine, err := NewEngine(opts, NewStaticQueryLimitsProvider(0), stats.NewQueryMetrics(nil), planner)
@@ -206,13 +207,23 @@ func TestUpstreamTestCases(t *testing.T) {
 }
 
 func TestOurTestCases(t *testing.T) {
-	opts := NewTestEngineOpts()
-	planner, err := NewQueryPlanner(opts)
-	require.NoError(t, err)
-	mimirEngine, err := NewEngine(opts, NewStaticQueryLimitsProvider(0), stats.NewQueryMetrics(nil), planner)
-	require.NoError(t, err)
+	makeEngines := func(t *testing.T, opts EngineOpts) (promql.QueryEngine, promql.QueryEngine) {
+		planner, err := NewQueryPlanner(opts)
+		require.NoError(t, err)
+		mimirEngine, err := NewEngine(opts, NewStaticQueryLimitsProvider(0), stats.NewQueryMetrics(nil), planner)
+		require.NoError(t, err)
 
-	prometheusEngine := promql.NewEngine(opts.CommonOpts)
+		prometheusEngine := promql.NewEngine(opts.CommonOpts)
+
+		return mimirEngine, prometheusEngine
+	}
+
+	opts := NewTestEngineOpts()
+	mimirEngine, prometheusEngine := makeEngines(t, opts)
+
+	optsWithDelayedNameRemoval := NewTestEngineOpts()
+	optsWithDelayedNameRemoval.CommonOpts.EnableDelayedNameRemoval = true
+	mimirEngineWithDelayedNameRemoval, prometheusEngineWithDelayedNameRemoval := makeEngines(t, optsWithDelayedNameRemoval)
 
 	testdataFS := os.DirFS("./testdata")
 	testFiles, err := fs.Glob(testdataFS, "ours*/*.test")
@@ -230,6 +241,11 @@ func TestOurTestCases(t *testing.T) {
 			testScript := string(b)
 
 			t.Run("Mimir's engine", func(t *testing.T) {
+				if strings.Contains(testFile, "name_label_dropping") {
+					promqltest.RunTest(t, testScript, mimirEngineWithDelayedNameRemoval)
+					return
+				}
+
 				promqltest.RunTest(t, testScript, mimirEngine)
 			})
 
@@ -237,6 +253,11 @@ func TestOurTestCases(t *testing.T) {
 			t.Run("Prometheus' engine", func(t *testing.T) {
 				if strings.HasPrefix(testFile, "ours-only") {
 					t.Skip("disabled for Prometheus' engine due to bug in Prometheus' engine")
+				}
+
+				if strings.Contains(testFile, "name_label_dropping") {
+					promqltest.RunTest(t, testScript, prometheusEngineWithDelayedNameRemoval)
+					return
 				}
 
 				promqltest.RunTest(t, testScript, prometheusEngine)